extract oracle logic into file
